scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
gspread>=5.12.0
oauth2client>=4.1.3
google-api-python-client>=2.100.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0

